Changed the default setting for the forever flag on display routines. Also we do not use the Pixel class anymore as it is not needed with the pre-processed data approach - simply a list of colours in the Display class is enough. The routines in the Display class that worked with the Pixel class have been moved to old.py.
